refactor: split configuration state tests
