+++ crates/mesh-llm-host-runtime/src/runtime/config_state_tests.rs
@@ -1,0 +1,23 @@
+use super::*;
+
+use crate::plugin::{GpuAssignment, GpuConfig, MeshConfig};
+use mesh_llm_config::{
+    ConfigDiagnosticCode, ConfigDiagnosticSeverity, validate_config_diagnostics,
+};
+use mesh_llm_plugin_manager::{
+    InstalledPluginConfigSchema, InstalledPluginManifestMetadata, InstalledPluginMetadata,
+    PluginStore, SUPPORTED_PLUGIN_SCHEMA_VERSION,
+};
+use std::collections::BTreeSet;
+
+#[path = "config_state_tests/support.rs"]
+mod support;
+use support::*;
+#[path = "config_state_tests/diagnostics.rs"]
+mod diagnostics;
+#[path = "config_state_tests/logging.rs"]
+mod logging;
+#[path = "config_state_tests/persistence.rs"]
+mod persistence;
+#[path = "config_state_tests/sync.rs"]
+mod sync;
